Clients: Move server test to common; Fix #4478
And use without pytest. Also remove rucio.tests from client package.
